Improve how worker creation and watching is handled
Use a timer to handle both the initial creation of workers as well as
periodically checking to see if a worker has died for whatever reason
and attempt to replace it. This approach solves two problems.

The first problem is when initially coming online, if workers could not
be started for whatever reason like the binary not existing for example,
execution would never make it into the event loop which prevented any
signal handling. This resulted in a situation where you could get stuck
in an infinite loop and not be able to stop the process without force
killing from an external source.

The second problem is previously when a worker died prematurely,
attempts were made to start a new one. If there was a failure starting a replacement worker
then there would be a small sleep before trying again in an infinite
fashion. This again meant possibly never returning to event loop. Now,
when a worker dies prematurely, it is accounted for and the timer will
detect this and attempt to start a new one. If that fails, it is logged
and will be tried again the next time the timer is invoked.

The timer is only enabled when the number of alive workers is not equal
to the number of requested workers. In the case where workers are all
present the timer is paused until it is need again.